remove a config import that is unused
this test checks that memoization works with all configs

test design: that's not necessarily true in future but it should be now.
in future perhaps a memoizer option would be "NoMemoizer" which would
not do *any* memoization, not even in-memory-only? but I think its ok
to not do that for now and i think its ok to require that a memoizer
always does actually do memoization at a memory level (so you can't
avoid it...)